fix:fix all the backend issues added api keys

+++ src/main/resources/db/migration/V4__create_cards_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE cards (
     back             TEXT NOT NULL,
     concept_name     VARCHAR(255) NOT NULL,
     concept_category VARCHAR(50) CHECK (concept_category IN ('definition','relationship','procedure','example','misconception')),
-    difficulty       SMALLINT CHECK (difficulty BETWEEN 1 AND 3),
+    difficulty       INTEGER CHECK (difficulty BETWEEN 1 AND 3),
     created_at       TIMESTAMP NOT NULL DEFAULT now()
 );
